pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "axm-genesis"
version = "1.2.0"
description = "Protocol for immutable, content-addressed knowledge graphs"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
  "blake3>=0.4.0",
  "pynacl>=1.5.0",
  "pyarrow>=14.0.0",
  "duckdb>=0.9.0",
  "click>=8.0.0",
]

[project.optional-dependencies]
pq = [
  "dilithium-py>=0.5.0",
]
dev = [
  "pytest>=7.0",
  "ruff>=0.1.0",
  "dilithium-py>=0.5.0",
]

[project.scripts]
axm-verify = "axm_verify.cli:main"
axm-build = "axm_build.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
